Move getEditorSetting to InterpreterFactory class
